Refactor live editing and add build
