+++ package.json
@@ -2,12 +2,12 @@
   "name": "igniteui-live-editing",
   "version": "1.0.0",
   "description": "This feature helps us to provide each sample as a separate application to external consumers.",
-  "main": "LiveEditingManager.js",
-  "types": "LiveEditingManager.d.js",
+  "main": "index.js",
+  "types": "index.d.ts",
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
-    "build": "rm -rf lib && npx tsc && cp package.json ./lib",
-    "build:dev": "tsc --watch"
+    "build": "gulp",
+    "build:dev": "gulp && gulp watch"
   },
   "repository": {
     "type": "git",
@@ -25,7 +25,9 @@
     "@babel/preset-env": "^7.12.11",
     "@babel/preset-typescript": "^7.12.7",
     "@types/node": "^14.14.19",
+    "del": "^6.0.0",
     "fs-extra": "^9.0.1",
+    "gulp": "^4.0.2",
     "slash": "^3.0.0",
     "ts-node": "^9.1.1",
     "typescript": "^4.1.3"
