+++ src/index.ts
@@ -1,0 +1,2 @@
+export * from './LiveEditingManager';
+export *  from './public';
