--- src/services/DependenciesType.ts
@@ -1,8 +1,0 @@
-export enum DependenciesType {
-    Default = 0,
-    Charts = 1,
-    Gauges = 2,
-    Excel = 3,
-    Maps = 4,
-    Spreadsheet = 5
-}
